+++ service/controller/resource/monitoring/pvcresizingresource/resource.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclient"
 	"github.com/giantswarm/micrologger"
+	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 )
 
 const (
@@ -12,19 +13,22 @@ const (
 )
 
 type Config struct {
-	K8sClient k8sclient.Interface
-	Logger    micrologger.Logger
+	K8sClient        k8sclient.Interface
+	PrometheusClient promclient.Interface
+	Logger           micrologger.Logger
 }
 
 type Resource struct {
-	k8sClient k8sclient.Interface
-	logger    micrologger.Logger
+	k8sClient  k8sclient.Interface
+	promClient promclient.Interface
+	logger     micrologger.Logger
 }
 
 func New(config Config) (*Resource, error) {
 	r := &Resource{
-		k8sClient: config.K8sClient,
-		logger:    config.Logger,
+		k8sClient:  config.K8sClient,
+		promClient: config.PrometheusClient,
+		logger:     config.Logger,
 	}
 
 	return r, nil